count articles in FAQ subcategories for explorer view
Currently visitors of the FAQ explorer might see the article count 0 if articles only exist
in subcategories. This might people make think that no articles exist at all. And additionally
the ticket overview by queue works differently.

To keep backwards compatibility the new behaviour has to be switched on in the sysconfig.